requirements.txt
selenium
requests
pillow
langchain
langchain-community
langchain-google-genai
google-generativeai
faiss-cpu
